#2180 - Clean up code related to project permission management
- Removing several oudated / deprecated / redundant methods
- Use modern API calls in some places
