+++ inception/inception-security/src/main/java/de/tudarmstadt/ukp/clarin/webanno/security/UserDao.java
@@ -17,11 +17,16 @@
  */
 package de.tudarmstadt.ukp.clarin.webanno.security;
 
+import static java.util.Arrays.asList;
+
 import java.util.List;
+import java.util.Properties;
 import java.util.Set;
 
 import de.tudarmstadt.ukp.clarin.webanno.security.model.Authority;
 import de.tudarmstadt.ukp.clarin.webanno.security.model.User;
+import de.tudarmstadt.ukp.clarin.webanno.support.ApplicationContextProvider;
+import de.tudarmstadt.ukp.clarin.webanno.support.SettingsUtil;
 
 /**
  * Provide methods for user management such as create, update, list users
@@ -128,4 +133,15 @@ public interface UserDao
      * Retrieve the number of enabled users
      */
     long countEnabledUsers();
+
+    public static boolean isProfileSelfServiceAllowed()
+    {
+        // If users are allowed to access their profile information, the also need to access the
+        // admin area. Note: access to the users own profile should be handled differently.
+        List<String> activeProfiles = asList(ApplicationContextProvider.getApplicationContext()
+                .getEnvironment().getActiveProfiles());
+        Properties settings = SettingsUtil.getSettings();
+        return !activeProfiles.contains("auto-mode-preauth")
+                && "true".equals(settings.getProperty(SettingsUtil.CFG_USER_ALLOW_PROFILE_ACCESS));
+    }
 }
